Organize test packages from core.api
